feat(mastra-multi-provider): implement AllowedModelRow and RefreshCatalogButton components for enhanced model management
