+++ apps/app/src/features/mastra/client/admin/AllowedModelsField.tsx
@@ -1,21 +1,15 @@
 import type { JSX } from 'react';
-import { useCallback, useId, useMemo, useState } from 'react';
+import { useCallback, useId, useMemo } from 'react';
 import { useTranslation } from 'next-i18next';
 import { useFieldArray, useFormContext, useWatch } from 'react-hook-form';
-import { Badge, Button, FormGroup, Input, Label } from 'reactstrap';
-
-import { ConfirmModal } from '~/client/components/Admin/App/ConfirmModal';
-import { apiv3Post } from '~/client/util/apiv3-client';
-import { toastError, toastSuccess } from '~/client/util/toastr';
+import { Button, FormGroup } from 'reactstrap';
 
 import type { AiProvider } from '../../interfaces/ai-provider';
-import type { RefreshModelCatalogResponse } from '../../interfaces/refresh-model-catalog-response';
-import { isValidProviderOptionsJson } from '../../utils/provider-options-validation';
+import { AllowedModelRow } from './AllowedModelRow';
 import type { AiSettingsFormValues } from './ai-settings-form-values';
 import { setDefaultAllowedModelAt } from './ai-settings-form-values';
-import { getProviderOptionsJsonStatus } from './provider-options-json-status';
 import { buildInitialProviderOptionsText } from './provider-options-namespace';
-import { registerToInputProps } from './register-to-input-props';
+import { RefreshCatalogButton } from './RefreshCatalogButton';
 import { useSWRxSelectableModels } from './use-selectable-models';
 
 // Vercel AI SDK docs describing the provider-namespaced `providerOptions` shape.
@@ -44,6 +38,8 @@ export interface AllowedModelsFieldProps {
  * The provider-scoped allowed-models editor, registered against the shared
  * react-hook-form context owned by `AiSettings`. One instance is mounted per
  * active provider panel (task 6.5 renders `<AllowedModelsField provider={p} />`).
+ * Each row is an `AllowedModelRow`; the models-section header carries a
+ * `RefreshCatalogButton`.
  *
  * ## index integrity (the load-bearing correctness rule)
  * A SINGLE `useFieldArray` spans the whole flat `allowedModels` array so the
@@ -132,51 +128,11 @@ export const AllowedModelsField = (
   }, [watchedModels, provider]);
 
   // Fetch the selectable models for THIS provider once at the field level and
-  // share the result with every row.
+  // share the result with every row. `invalidateAllProviders` is provider-
+  // independent (the refresh replaces the whole models.dev snapshot server-side),
+  // so it is handed to the refresh button.
   const { data, error, invalidateAllProviders } =
     useSWRxSelectableModels(provider);
-
-  // Manual catalog refresh: ask the server to re-ingest models.dev and persist
-  // the snapshot, then invalidate every cached provider list — the snapshot is
-  // replaced for ALL providers server-side, so a mutate bound to the current
-  // provider only would leave other visited providers' immutable caches
-  // pre-refresh until a page reload. Preserved unchanged through the
-  // provider-scoping (it is picker-owned functionality). Intentionally NOT
-  // disabled in env-only mode: the catalog is a server-side cache of public
-  // model metadata, not an AI setting, and env-only deployments (e.g.
-  // GROWI.cloud) are a primary audience of this action.
-  const [isRefreshingCatalog, setRefreshingCatalog] = useState(false);
-  const refreshCatalog = useCallback(async (): Promise<void> => {
-    setRefreshingCatalog(true);
-    try {
-      await apiv3Post<RefreshModelCatalogResponse>(
-        '/ai-settings/refresh-model-catalog',
-      );
-      await invalidateAllProviders();
-      toastSuccess(t('ai_settings.refresh_model_catalog_success'));
-    } catch {
-      // The server answers a generic 500 on failure (the last-good catalog
-      // stays in effect) — surface the localized failure message instead.
-      toastError(t('ai_settings.refresh_model_catalog_failed'));
-    } finally {
-      setRefreshingCatalog(false);
-    }
-  }, [invalidateAllProviders, t]);
-
-  // The refresh triggers server-side OUTBOUND communication (models.dev), so
-  // the button opens a confirmation first — the request runs only after the
-  // admin explicitly confirms.
-  const [isRefreshConfirmOpen, setRefreshConfirmOpen] = useState(false);
-  const openRefreshConfirm = useCallback((): void => {
-    setRefreshConfirmOpen(true);
-  }, []);
-  const cancelRefreshConfirm = useCallback((): void => {
-    setRefreshConfirmOpen(false);
-  }, []);
-  const confirmRefreshCatalog = useCallback(async (): Promise<void> => {
-    setRefreshConfirmOpen(false);
-    await refreshCatalog();
-  }, [refreshCatalog]);
 
   // Mode derivation:
   // - `select` only when the catalog resolved to a non-empty list (R2.6).
@@ -268,31 +224,8 @@ export const AllowedModelsField = (
         <h3 className="fs-5 fw-bold m-0">
           {t('ai_settings.models_section_title')}
         </h3>
-        <Button
-          type="button"
-          color="link"
-          size="sm"
-          className="ms-auto p-0 d-inline-flex align-items-center"
-          disabled={isRefreshingCatalog}
-          onClick={openRefreshConfirm}
-        >
-          <span
-            className="material-symbols-outlined fs-6 me-1"
-            aria-hidden="true"
-          >
-            refresh
-          </span>
-          {t('ai_settings.refresh_model_catalog')}
-        </Button>
+        <RefreshCatalogButton invalidateAllProviders={invalidateAllProviders} />
       </div>
-      <ConfirmModal
-        isModalOpen={isRefreshConfirmOpen}
-        warningMessage={t('ai_settings.refresh_model_catalog_confirmation')}
-        supplymentaryMessage={null}
-        confirmButtonTitle={t('ai_settings.refresh_model_catalog_confirm')}
-        onConfirm={confirmRefreshCatalog}
-        onCancel={cancelRefreshConfirm}
-      />
       <p className="form-text text-muted mt-0 mb-3">
         {t('ai_settings.models_section_desc')}
       </p>
@@ -339,249 +272,6 @@ export const AllowedModelsField = (
         </span>
         {t(addLabelKey)}
       </Button>
-    </FormGroup>
-  );
-};
-
-interface AllowedModelRowProps {
-  /** Position of this row in the flat `allowedModels` array (NOT the display index). */
-  readonly originalIndex: number;
-  /**
-   * Whether this row is the global default. Provided by the parent from its
-   * whole-array watch (not a row-local nested watch) so it reflects the array-root
-   * `setValue` used by the ★ pick and the delete-reassign.
-   */
-  readonly isDefault: boolean;
-  readonly labelKey: string;
-  readonly radioGroupName: string;
-  readonly disabled: boolean;
-  /**
-   * Render the modelId control as a select-only dropdown (`true`) when the
-   * provider has a non-empty catalog, or as free-text input (`false`) otherwise.
-   */
-  readonly useSelect: boolean;
-  /** The catalog model ids offered as dropdown options (empty in free-text mode). */
-  readonly selectableModelIds: readonly string[];
-  /** Non-empty model ids already registered under this provider (any row). */
-  readonly registeredModelIds: ReadonlySet<string>;
-  /** Model ids duplicated within this provider — drives the row-level error. */
-  readonly duplicateModelIds: ReadonlySet<string>;
-  /** The catalog fetch is in flight; the modelId control is disabled meanwhile. */
-  readonly isLoadingModels: boolean;
-  readonly docUrl: string;
-  readonly placeholder: string;
-  readonly onSelectDefault: () => void;
-  readonly onRemove: () => void;
-}
-
-/**
- * One allowed-model card: model id (monospace) + "default" badge/radio + remove
- * trash icon + providerOptions JSON with a live valid/invalid indicator, a format
- * link, and a docs link. Extracted so each card owns its own field ids and
- * watches only its own fields (isDefault + modelId + providerOptions value). All
- * register/watch paths are keyed on `originalIndex` (the flat-array position).
- */
-const AllowedModelRow = (props: AllowedModelRowProps): JSX.Element => {
-  const {
-    originalIndex,
-    isDefault,
-    labelKey,
-    radioGroupName,
-    disabled,
-    useSelect,
-    selectableModelIds,
-    registeredModelIds,
-    duplicateModelIds,
-    isLoadingModels,
-    docUrl,
-    placeholder,
-    onSelectDefault,
-    onRemove,
-  } = props;
-  const { t } = useTranslation('admin');
-  const { control, register } = useFormContext<AiSettingsFormValues>();
-
-  const modelInputId = useId();
-  const providerOptionsId = useId();
-  const radioId = useId();
-
-  // Watch only this card's own value fields (modelId + providerOptions) so
-  // editing a row re-renders just that row. `isDefault` comes from the parent
-  // (see AllowedModelRowProps) — it is set by an array-root rewrite.
-  const providerOptionsText =
-    useWatch({
-      control,
-      name: `allowedModels.${originalIndex}.providerOptionsText`,
-    }) ?? '';
-  const currentModelId =
-    useWatch({ control, name: `allowedModels.${originalIndex}.modelId` }) ?? '';
-
-  // Registered-excluded options (R2.6): offer catalog ids NOT already registered
-  // by another row of this provider, but always keep this row's OWN current value
-  // selectable (so switching this row's model is possible and its saved value is
-  // never dropped).
-  const availableModelIds = selectableModelIds.filter(
-    (id) => id === currentModelId || !registeredModelIds.has(id),
-  );
-  // A saved value absent from the current catalog is preserved as its own option
-  // so it is neither reset nor silently changed.
-  const hasOutOfListValue =
-    currentModelId !== '' && !selectableModelIds.includes(currentModelId);
-
-  // Same-provider duplicate (R2.4): flagged when this row's non-empty id collides
-  // with another row of the same provider.
-  const isDuplicate =
-    currentModelId !== '' && duplicateModelIds.has(currentModelId);
-
-  const status = useMemo(
-    () => getProviderOptionsJsonStatus(providerOptionsText),
-    [providerOptionsText],
-  );
-  const isInvalidJson =
-    status.kind === 'syntax-error' || status.kind === 'shape-error';
-
-  return (
-    <FormGroup
-      tag="fieldset"
-      className="rounded p-3 mb-2 border"
-      data-testid="allowed-model-row"
-    >
-      {/* The label/badge sit on their own line; the input, default radio, and
-          remove icon share one center-aligned row below. */}
-      <div className="mb-2">
-        <div className="d-flex align-items-center gap-2 mb-1">
-          <Label for={modelInputId} className="form-label small mb-0">
-            {t(labelKey)}
-          </Label>
-          {isDefault && (
-            <Badge color="info" pill>
-              {t('ai_settings.default_badge')}
-            </Badge>
-          )}
-        </div>
-        <div className="d-flex align-items-center gap-3">
-          {/* The form binding (`register(...modelId)`) and value format are
-              identical in both modes; only the control type differs — a
-              select-only dropdown when the provider has a catalog (R2.6),
-              otherwise the free-text input (catalog-less provider / fetch failure
-              — R2.7). Disabled in env-only mode and while the catalog is loading. */}
-          <Input
-            id={modelInputId}
-            type={useSelect ? 'select' : 'text'}
-            className="font-monospace flex-grow-1"
-            disabled={disabled || isLoadingModels}
-            invalid={isDuplicate}
-            {...registerToInputProps(
-              register(`allowedModels.${originalIndex}.modelId`),
-            )}
-          >
-            {/* Options only exist in select mode. Free-text mode must pass
-                `undefined` (NOT `false`): a text <input> is a void element, and
-                React rejects any non-null child — reactstrap only strips a
-                *truthy* child, so `false` would crash. */}
-            {useSelect ? (
-              <>
-                <option value="">{t('ai_settings.model_placeholder')}</option>
-                {availableModelIds.map((id) => (
-                  <option key={id} value={id}>
-                    {id}
-                  </option>
-                ))}
-                {hasOutOfListValue && (
-                  <option value={currentModelId}>{currentModelId}</option>
-                )}
-              </>
-            ) : undefined}
-          </Input>
-          <FormGroup check className="mb-0 text-nowrap">
-            <Input
-              id={radioId}
-              type="radio"
-              role="radio"
-              name={radioGroupName}
-              checked={isDefault}
-              disabled={disabled}
-              onChange={onSelectDefault}
-            />
-            <Label check for={radioId} className="ms-1">
-              {t('ai_settings.set_as_default')}
-            </Label>
-          </FormGroup>
-          <Button
-            type="button"
-            color="link"
-            className="text-body-secondary p-1"
-            aria-label={t('ai_settings.remove_model')}
-            title={t('ai_settings.remove_model')}
-            disabled={disabled}
-            onClick={onRemove}
-          >
-            <span className="material-symbols-outlined" aria-hidden="true">
-              delete
-            </span>
-          </Button>
-        </div>
-        {isDuplicate && (
-          <div className="invalid-feedback d-block">
-            {t('ai_settings.model_duplicate_error')}
-          </div>
-        )}
-      </div>
-
-      <div>
-        <div className="d-flex align-items-center mb-1">
-          <Label for={providerOptionsId} className="form-label small mb-0">
-            {t('ai_settings.provider_options_label')}
-          </Label>
-          <a
-            href={docUrl}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="ms-auto small d-inline-flex align-items-center"
-          >
-            {t('ai_settings.provider_options_doc_link')}
-            <span
-              className="material-symbols-outlined fs-6 ms-1"
-              aria-hidden="true"
-            >
-              open_in_new
-            </span>
-          </a>
-        </div>
-        {/* Suppress Bootstrap's `.is-invalid` background icon: on a textarea it
-            sits at the top-right and gets clipped by the scrollbar once the
-            content overflows. The red border + the message below convey the
-            invalid state without it. */}
-        <Input
-          id={providerOptionsId}
-          type="textarea"
-          rows={6}
-          className="font-monospace"
-          placeholder={placeholder}
-          disabled={disabled}
-          invalid={isInvalidJson}
-          style={{ backgroundImage: 'none' }}
-          {...registerToInputProps(
-            register(`allowedModels.${originalIndex}.providerOptionsText`, {
-              validate: (v) =>
-                isValidProviderOptionsJson(v) ||
-                t('ai_settings.provider_options_invalid_json'),
-            }),
-          )}
-        />
-        {isInvalidJson && (
-          <div className="invalid-feedback d-block">
-            {t('ai_settings.provider_options_invalid_json')}
-            {status.kind === 'syntax-error' && (
-              <span className="ms-1">
-                {t('ai_settings.provider_options_error_at_line', {
-                  line: status.line,
-                })}
-              </span>
-            )}
-          </div>
-        )}
-      </div>
     </FormGroup>
   );
 };
